Double the resolution
Instead of using cells that are 2x2 tiles, use cells that are 1x1 tiles.
This change required rewriting all of the cell indexing and lookup
routines as well as the PPU update method.  No longer can we use the
attribute table to update the whole screen at once.  We must now update
each tile individually.  Rows are processed one at a time and are
buffered & written to the PPU every frame.

Each cell is now represented by a single bit instead of a full byte.  A
bitmask is used to inspect and update individual cells.